// frontend/src/api/client/models/Fahrzeugklasse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 */
export const Fahrzeugklasse = {
  Rad: "RAD",
  SummeKfz: "SUMME_KFZ",
  ZweiPlusEins: "ZWEI_PLUS_EINS",
  AchtPlusEins: "ACHT_PLUS_EINS",
} as const;
export type Fahrzeugklasse =
  (typeof Fahrzeugklasse)[keyof typeof Fahrzeugklasse];

export function instanceOfFahrzeugklasse(value: any): boolean {
  for (const key in Fahrzeugklasse) {
    if (Object.prototype.hasOwnProperty.call(Fahrzeugklasse, key)) {
      if (Fahrzeugklasse[key as keyof typeof Fahrzeugklasse] === value) {
        return true;
      }
    }
  }
  return false;
}

export function FahrzeugklasseFromJSON(json: any): Fahrzeugklasse {
  return FahrzeugklasseFromJSONTyped(json, false);
}

export function FahrzeugklasseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): Fahrzeugklasse {
  return json as Fahrzeugklasse;
}

export function FahrzeugklasseToJSON(value?: Fahrzeugklasse | null): any {
  return value as any;
}

export function FahrzeugklasseToJSONTyped(
  value: any,
  ignoreDiscriminator: boolean
): Fahrzeugklasse {
  return value as Fahrzeugklasse;
}