// frontend/src/api/client/models/ZaehldatenIntervall.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 */
export const ZaehldatenIntervall = {
  StundeViertel: "STUNDE_VIERTEL",
  StundeViertelEingeschraenkt: "STUNDE_VIERTEL_EINGESCHRAENKT",
  StundeHalb: "STUNDE_HALB",
  StundeKomplett: "STUNDE_KOMPLETT",
} as const;
export type ZaehldatenIntervall =
  (typeof ZaehldatenIntervall)[keyof typeof ZaehldatenIntervall];

export function instanceOfZaehldatenIntervall(value: any): boolean {
  for (const key in ZaehldatenIntervall) {
    if (Object.prototype.hasOwnProperty.call(ZaehldatenIntervall, key)) {
      if (
        ZaehldatenIntervall[key as keyof typeof ZaehldatenIntervall] === value
      ) {
        return true;
      }
    }
  }
  return false;
}

export function ZaehldatenIntervallFromJSON(json: any): ZaehldatenIntervall {
  return ZaehldatenIntervallFromJSONTyped(json, false);
}

export function ZaehldatenIntervallFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean
): ZaehldatenIntervall {
  return json as ZaehldatenIntervall;
}

export function ZaehldatenIntervallToJSON(
  value?: ZaehldatenIntervall | null
): any {
  return value as any;
}

export function ZaehldatenIntervallToJSONTyped(
  value: any,
  ignoreDiscriminator: boolean
): ZaehldatenIntervall {
  return value as ZaehldatenIntervall;
}